+++ backend/index.ts
@@ -36,7 +36,13 @@ const formidableHandle = (name, req, res, handle) =>
 export default {
 	"/musicxml2ly": {
 		post: (req, res) => formidableHandle("musicxml2ly", req, res,
-			({xml, options}) => lilyCommands.xml2ly(xml, options && JSON.stringify(options))),
+			({xml, options}) => lilyCommands.xml2ly(xml, options && JSON.parse(options))),
+	},
+
+
+	"/midi2ly": {
+		post: (req, res) => formidableHandle("midi2ly", req, res,
+			({options}, {midi}) => lilyCommands.midi2ly(midi, options && JSON.parse(options))),
 	},